Refactor clusters-kmeans
