+++ packages/turf-clusters-kmeans/.gitignore
@@ -1,0 +1,1 @@
+index.js
