+++ packages/turf-clusters-kmeans/bench.js
@@ -1,8 +1,8 @@
-import fs from 'fs';
-import path from 'path';
-import load from 'load-json-file';
-import Benchmark from 'benchmark';
-import clustersKmeans from './';
+const fs = require('fs');
+const path = require('path');
+const load = require('load-json-file');
+const Benchmark = require('benchmark');
+const clustersKmeans = require('./').default;
 
 // Define Fixtures
 const directory = path.join(__dirname, 'test', 'in') + path.sep;
